+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 0.52.1
  June 30, 2021
 - _con.initModule(_tabController.index);
+- addState(subModuleStates[moduleType + module['name']]);
 
 ## 0.52.0
  June 29, 2021